bootstrap:topdeps: give newtype "pretend wrapper" members topdep to real member
	newtype c = b

In c, autointf creates "pretend wrappers" that correspond to the member
functions and methods of b. Cprinter actually skips those and we use the
members of b directly. In fact, all occurrences of c are simply replaced
with b. To make sure the needed members of b are printed by cprinter, we
add a dependency to them.