bootstrap:lir: insert Trace_except calls
Trace_except is very simple for now: it prints to stderr whenever the
error is not OK. This is noisy, e.g. when recoverable errors are
generated.

Future plans include writing messages to a global circular buffer, and
only print the except "backtrace" on hitting a "Fatal", ... as opposed
to a "Drop".