n.strconv: return Error second since the value carries information
The convention in N (and probably to be enforced by the compiler at
some point) is that any return values after a non-OK error must be
ignored, and in fact must be zeroed.

If we're going to return some information, even in case of error, then
the value must come first.