bootstrap: support vararg forwarding with extra arguments
Until now, it was only possible to pass varargs down if it was the only
(variable) argument to the (second) vararg function. It was often
possible to work around it by doing:

	fun prefix_print ...v:(Vararg *`Show)
		fmt.Prf "prefix"
		fmt.Prf ...v

Now, you can do:

	fun prefix_print ...v:(Vararg *`Show)
		fmt.Prf "prefix" ...v

Similarly, it is now also possible to pass in a slice in a variable
argument position with other variable arguments before it.

This is implemented by adding high-bit flags to the vacount argument:
these bits are checked when getting the last argument of a vararg.

Also, ensure that local variables declared in macros use N reserved
names.