bootstrap: drop, fatal, never: except variants
Can be used where except can. Instead of throwing the error if it is
non-OK, they drop it or abort the program.

	drop = file#Close
	never = fmt.F stringbuf "test"
	let fatal, rnd = fs.Open "/dev/random"

fatal and never convey a different semantic meaning even though they
behave the same right now. The semantic system would complain if it can
prove that a never can be triggered. A fatal, by definition, can be
triggered -- one just hopes it doesn't.