n.net.http: Server, `Server, and `Handler
A Server creates and listen on a socket. On accepting a connection, it
creates a conn_handler which will be responsible for the connection over
its lifetime (multiple request/response may go over the connection).

Once the header is available, the conn_handler looks for a `Server
(among those registered with the Server) willing to process the request.
The `Server returns a `Handler capable of processing the request and
generating a response.

A `Handler is then called whenever an event happens on the underlying
socket (by now the Header is fully loaded). A handler can start
doing work on a partial request body (e.g. large upload). It writes the
response incrementally (but it must read the whole request first, as per
HTTP/1.x).

Handlers must be non-blocking.

It will be easy to create simplified handlers on top (for instance,
handlers that are always presented with a full header and a full request
body; or handlers that can return a whole response at once -- a buffer,
or a file -- and delegate the work of writing that content in a
non-blocking manner).