bootstrap:cprinter: if no ssa rewrite of "x = INIT", we must zero before assign
We used to rewrite "x = INIT" into

	let gen = INIT
	x = gen

Now we keep "x = INIT". But such assignments are turned into a simple
INIT with "x" as the INIT.target_expr, and are codegened as a series of
statements:

	x.field1 = val1
	x.field2 = val2

So we first need to make sure x is fully zeroed. We now generate the
statement:

	x = (type){ 0 };