utils:debugedit: rewrite .debug_line using ncc0 linemap
By running debugedit on binaries created by ncc0, we can now use gdb to
step through the N code.

	./ncc0 t00/arith.n
	make -C utils/debugedit
	utils/debugedit/debugedit a.out
	gdb a.out

We will likely run this as part of the normal compilation process soon.

The rewriting process is quite tricky. DWARF isn't a nice format.

Starting from RPM's debugedit seemed logical, but in hindsight it would
have been easier to write this from scratch. debugedit is not nice code,
and N is a better language for this kind of byte shuffling anyway.