bootstrap:lexer: escaped lines extend an single-line block with a single tab
Previously, it was necessary to write

	assert false
			\ or true

Now we can write the more consistent

	assert false
		\ or true

The rules for extending single-line blocks are still weird (see
t00/block.n for a weird case). We have plans to overall single-line,
multiple-statement blocks (using {} and single ;), but we still want
single-line (implicit) single-statement blocks (for assert, pre, etc.).
It's unclear how this will work.