n.net.http: client support
Implemented to be used in an non-blocking event loop. Simpler usage
patterns will be possible on top.

We had misunderstood the role played by the EPOLLHUP event (which
indicates connection hangup, not that the file descriptor was closed by
someone else, e.g. the kernel, as we thought we read somewhere), so we
introduce a special CLOSE event to be used instead of HUP to request
that ev.Poll descriptor at the appropriate time.