n.fs: File.Open_fd; delegate EOF generation to syscall.Read
The test is tricky enough that we should do it at the lowest level: if
Available capacity in the buffer was 0, and we read 0 bytes, it doesn't
mean io.EOF.

Our decision to express syscall.Read in terms of slices means we should
also take of EOF-level of details.

It is conceivable that a caller may want to see through such
abstractions, but they are free to write their own read(2) wrapper.