n.ev: add ability to trigger a handler manually
We don't want to call the handler directly, we want to schedule it to be
called from within the event processing loop.

We have a special event type MANUAL: we don't want to encourage the
creation of fake I/O events, which could very easily break the rules of
epoll/EAGAIN on non-blocking devices.