n.syscall.Fd: manage file descriptor lifetimes
Manual management remains the rule, you must call syscall.Close
explicitly. The goal is to prevent mistakes and ensure that a fd value
is not used after it is closed (dangling fd), and to prevent fds that
are never closed (unreachable fd).

Closing fds when the last reference is destroyed is questionable: after
all, a fd may have a well-known integer value (e.g. 2) and may not
require storage in the program. But we consider this to be a
corner case.