n.builtins.`Any_slice: Unsafe_data, Unsafe_rawdata: return nullable ref
When the slice is empty, we have nothing else to return than nil.
Previously, we asserted that the slice wasn't empty, but this was
misguided. Instead of crashing in an non-easy to expect way (without
realizing that if these functions do not return nullable refs, then they
must be prone to crashing... and in the absence of inline pre), simply
return nullable references and force the caller to handle the nil case.
Besides, many users of these methods interact with C libraries, and most
such code will (hopefully) be capable of handling NULL.