n.fs.File.Open_fd: File should be exposed via references
Primarily for consistency with fs.Open et al.

This shouldn't actually matter, but there is a codegen bug that affects
us in nastier ways if we let files be allocated on the stack / within a
struct, namely the lack of automatic dtor of struct fields if there is
an explicit Dtor definition. We will fix File.Dtor in the next patch to
workaround ncc0 stupid codegen. This is only incidental but made the
discrepancy with fs.Open et al. evident.