n.bufio.`Fillable_reader.Fill: also return count of bytes just read
	met# Fill ?max:?Uint = just_filled:Uint, Error

When stacking multiple buffers, for instance with some of them
performing decoding, it is useful to know if *more* data has just been
read: data may already have been Buffered before, but maybe not
enough was available to progress with decoding.

It's also helpful to have just_filled available when considering the
error. If just_filled is > 0, we may be able to use the data just read
right away, and the error may possibly be left for later consideration,
or discarded, waiting to see what a next call to Fill will say.