n.syscall.Write, Pwrite: return io.SHORT_WRITE on short write and errno==0
POSIX says an error may be detected on a short write, but not
necessarily. io.`Write says short writes must return an error. Handling
such details in the syscall module is debatable, we do here as with
syscall.Read and io.EOF.

The idea of requiring short writes to return an error comes from Go's
io.Write interface. And indeed, in os/file_unix.go File.Write() retries
after short writes (as well as after EINTR). But Go assumes
application-level parallelism with goroutines and can afford to
block multiple times in a row on a syscall, whereas in N we want to
support non-blocking i/o and predictable "blocking" points. Therefore,
we may want to reconsider this requirement on short writes in
io.`Writer.