bootstrap:parser: switch to "for(each)? ... over ..."
Because of our support for destruct such as:

	for x, y in iterator
	foreach x, y in iterator

and to keep things simple and consistent, we had TCOMMA with a higher
precedence than Tin. But to keep boolean expression sane, and because
Tin is also used to express set membership, we ended up with the
precedence hiearchy:

	Telse < Tor < Tand < Tnot < Tin < Tcomma

which allowed writing

	x in {0 1} or y

without parentheses.

The problem was with constructing tuples. While

	return 0:U32, x==y

returned a (U32, Bool)

	return 0:U32, x in {0 1}

is a type error, as the tuple (0:U32, x) cannot be a member of that set
literal.

So we introduce "over" for use in loop constructs (thanks Jess!), with
the saner precedences:

	Tover < Tcomma < Telse < Tor < Tand < Tnot < Tin