boostrap:cprinter: fix pass-by-value for non-trivial types
For arguments passed by-value, the caller makes a copy and passes it in
to the callee, which becomes responsible for it.

These two principles must apply:
- callee must dtor if non-`Trivial_dtor;
- pass throug a ref if non-`Trivial_move.

To keep things simple, we reuse `Return_by_copy (which we could rename
to pass-by-copy). Anything non-`Return_by_copy that is passed by-value
will be dtor by the callee and passed through a ref in C.