bootstrap: infer (symbol) export for functions used by inline functions
When an inline function f uses a non-exported, non-inline function g, we
used to generate g as a static function in all the compilation units
that needed it.

The correct approach for the N ABI model is to expose the symbol of
the function g to the other compilations units that need to call it (and
declare their prototype), but keep the function's body private. We set
the toplevel->flags TOP_IS_FORCE_SYMBOL_EXPORT on functions like g.

Note that for global LET, we still export (the body of) the definition
in other compilation units. TODO