n.encoding.protobuf.Reader: return io.EOF when there was nothing at all
We used to return io.UNEXPECTED_EOF even when no value was found at all,
whereas this error makes more sense when eof happens in the middle of
decoding a value.

Makes it possible to find EOF on a stream simply by trying to read an
encoded value off of it.