n.encoding.protobuf.Reader.Catchup_reader: sync underlying reader
protobuf.Reader is buffered. Catchup_reader syncs up the underlying
reader as far as the decoder has gone. This is useful when the byte
position of protobuf values in the byte stream is of interest.

Also, add a Copy_to:@#io.`Writer field to protobuf.Reader. When set,
copy all encoded bytes consumed to it.