bootstrap: node_move_content: fix: properly cleanup internal typ fields
So far, only DIRECTDEF nodes had an internal field other than .typ
(namely as.DIRECTDEF.typ) and we (apparently) never moved the content of
such nodes.

Now we have CONV nodes that we do move, and they have an internal field
as.CONV.to. Care must be taken to move such a field properly (in
particular, any backlinks must be removed).

We use node_set_which(src, 0) and node_set_which(dst, 0) for this
purpose.