n: String and Buf: let Reserve shrink memory; use auto_grow more
String, and to a lesser extent Buf, were using Reserve to allocate more
capacity in the implementation of multiple of their methods, however
Reserve is exact (in the current implementation): it only allocates the
requested capacity, never more.

We replace most of these uses with auto_grow, which uses
Buffer_auto_grown_size which follows an exponential-lite growth
strategy.

Additionally, Reserve will now shrink the capacity if the requested size
is less than the current capacity and more than the current length.