n.builtins: register _$Npostlude with atexit(3) and add n.builtins.Exit
We previously only invoked the postlude in main(), passing it the return
value of _$Nmain (aka. Main). Since we didn't really have an exit
function that wasn't so bad (syscall.Exit is _exit(2) which doesn't call
atexit-registered functions).

We introduce n.builtins.Exit, which calls exit(3) but also stores its
argument in a global variable so _$Npostlude can access it.

The prelude registers the postlude with atexit(3).

main() stores the return value of _$Nmain in the global variable.